Works fine by now.

+++ src/main/java/net/ddns/worldofjarcraft/DatabaseRepresentation/EinkaufRepository.java
@@ -6,5 +6,5 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface EinkaufRepository extends CrudRepository<Einkauf, Long> {
+public interface EinkaufRepository extends CrudRepository<Einkauf, Integer> {
 }
